+++ extensions/android/ringqt/project/ring/include/vm.h
@@ -633,7 +633,7 @@
 
 	void ring_vm_not ( VM *pVM ) ;
 
-	double ring_vm_stringtologicvalue ( VM *pVM,const char *cStr ) ;
+	int ring_vm_stringtologicvalue ( VM *pVM,const char *cStr ) ;
 
 	int ring_vm_listtologicvalue ( VM *pVM,List *pList ) ;
 	/* Variables */
